--- server/src/main/java/com/codestates/server/global/security/oauth2/v4/HelloHomeController.java
@@ -1,25 +1,0 @@
-package com.codestates.server.global.security.oauth2.v4;
-
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
-import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
-import org.springframework.security.oauth2.core.OAuth2AccessToken;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HelloHomeController {
-
-    @GetMapping("/oauth2")
-    public String home(@RegisteredOAuth2AuthorizedClient("kakao")OAuth2AuthorizedClient authorizedClient) {
-
-        OAuth2AccessToken accessToken = authorizedClient.getAccessToken();
-        System.out.println("Access Token Value: " + accessToken.getTokenValue());
-        System.out.println("Access Token Type: " + accessToken.getTokenType().getValue());
-        System.out.println("Access Token Scopes: " + accessToken.getScopes());
-        System.out.println("Access Token Issued At: " + accessToken.getIssuedAt());
-        System.out.println("Access Token Expires At: " + accessToken.getExpiresAt());
-
-        return "hello-oauth";
-
-    }
-}
